# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for live stream configuration, ingest/publish lifecycle, DVR recordings, and event hooks.",
        "title": "Open Streamer API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/watermarks": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watermarks"
                ],
                "summary": "List watermark assets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.WatermarkAssetList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watermarks"
                ],
                "summary": "Upload a watermark image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image file (PNG / JPG / GIF, ≤ 8 MiB)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Display name (defaults to filename)",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/apidocs.WatermarkAssetData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "413": {
                        "description": "Request Entity Too Large",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "415": {
                        "description": "Unsupported Media Type",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            }
        },
        "/watermarks/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "watermarks"
                ],
                "summary": "Get watermark metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.WatermarkAssetData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "watermarks"
                ],
                "summary": "Delete a watermark asset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            }
        },
        "/watermarks/{id}/raw": {
            "get": {
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "watermarks"
                ],
                "summary": "Download watermark image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "apidocs.ErrorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "apidocs.WatermarkAssetData": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/domain.WatermarkAsset"
                }
            }
        },
        "apidocs.WatermarkAssetList": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.WatermarkAsset"
                    }
                },
                "dir": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "domain.WatermarkAsset": {
            "type": "object",
            "properties": {
                "content_type": {
                    "description": "ContentType is the MIME type sniffed at upload time\n(e.g. \"image/png\", \"image/jpeg\"). Used for Content-Type on /raw GET.",
                    "type": "string"
                },
                "file_name": {
                    "description": "FileName is the original filename at upload time, preserved for\naudit / download responses. Includes extension (e.g. \"logo.png\").",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the immutable filesystem-safe identifier (also the on-disk basename).",
                    "type": "string"
                },
                "name": {
                    "description": "Name is the human display label shown in UIs. Defaults to the upload's\noriginal filename when the operator doesn't override.",
                    "type": "string"
                },
                "size_bytes": {
                    "description": "SizeBytes is the on-disk size of the image. Sidecar JSON is excluded.",
                    "type": "integer"
                },
                "uploaded_at": {
                    "description": "UploadedAt is the wall-clock UTC time the asset was first stored.",
                    "type": "string"
                }
            }
        },
...
    }
}